Italian Language File
Hi, my name is Alberto Di Mattei. For years I use php4dvd and I have always updated the Italian language file by myself. Since no one seems to send it to you I do it. The file is related to version 3.6.2. If you want to enter me in the contributors list, I'll be happy to update it periodically. Bye and thank you for your work!